refactor: fix responsive styles across platforms
