+++ packages/key-facts/src/key-facts-title.web.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { propTypes, defaultProps } from "./key-facts-title-prop-types";
+import { KeyFactsTitleResponsive } from "./styles/responsive";
+import styles from "./styles";
+
+const KeyFactsTitle = ({ title }) => (
+  <KeyFactsTitleResponsive style={styles.title}>
+    {title.toUpperCase()}
+  </KeyFactsTitleResponsive>
+);
+
+KeyFactsTitle.propTypes = propTypes;
+KeyFactsTitle.defaultProps = defaultProps;
+
+export default KeyFactsTitle;
