--- packages/key-facts/src/styles/index.web.js
@@ -1,14 +1,0 @@
-import { StyleSheet } from "react-native";
-import { spacing } from "@times-components/styleguide";
-import sharedStyles from "./shared";
-
-const styles = StyleSheet.create({
-  ...sharedStyles,
-  title: {
-    ...sharedStyles.title,
-    minWidth: 175,
-    paddingRight: spacing(4)
-  }
-});
-
-export default styles;
